requirements.txt
gradio
openai
pandas
matplotlib
python-dotenv
